Fixed a bug on the search bar.
Edited appearance of the top navbar. (+ other styles parts)
The navbar's height is no longer a fixed value. Although the spacing element still needs a value, since it's a fixed element.

Edited popup messages (now they don't look as ugly, and they appear from the bottom, doing a fadein and fadeout effect)

Updated the nightmode's background so it's the same as lightmode, but darker. (Instead of ugly bricks)

Translated the comments and JSON responses to English.

Hid the discover module so it's no longer accessible.

Added some explanations to some functions.

Format...