#41 Reworked how alerts work
#41 Adapted to every part of the code. Now alerts can be either permanent, or temporal, stacking.

Now the children emit to the parent, so the alerts display on the parent (to avoid having to add timeouts before redirecting, because otherwise alerts would disappear as you switched to another component, with a not too smooth animation).

Also changed the way emit worked, now it emits an object with attributes that define what it does, instead of just a string/language object. Mapped object defined in struct.ts. (Maps the emitter types)

Changed currentLang type to 'any', due to some problems. Defining an interface could be an option as well for some of the variables where I used 'any', but prob not for the language object.