#43  Added modals for following/followed
Added styles for modal in night mode as well.

Changed the message when registering, and autoredirecting (it used to need a click after registering, which was worse than a simple redirect imo)

Corrected the placeholder when editting the user bio.

Renamed a function in PHP for clarity, due to the function now returning an object of followers and following.

Changed/added a bunch of styles to improve the view.