Fixed problem with redirecting to upload image (when logged in)
Added way to close maximized view.

Added condition to only save last URL if it's not the same as the current one (reloading same page).

Fixed zoom var in scrolling events. It was 0.8, as CSS in the past. It no longer uses zoom, as indicated in #44.

Added way to center (vertically) modals, adding class when creating it. See app.component.ts > fn open.

Pointed errors to solve (already stated on issue #48).

General styling.